requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
openai>=1.0.0
pyyaml>=6.0
psutil>=5.9.0
streamlit>=1.28.0
Pillow>=10.0.0

